+++ packages/twenty-website/src/app/components/ContentContainer.tsx
@@ -5,9 +5,11 @@ import styled from '@emotion/styled'
 const Container = styled.div`
     display: flex;
     flex-direction: column;
-    width: 600px;
+    width: 100%;
+    padding: 0px 96px 0px 96px;
      @media(max-width: 809px) {
         width: 100%;
+        padding: 0px 12px 0px 12px;
     }`;
