+++ packages/twenty-website/src/content/user-guide/basics/custom-objects.mdx
@@ -1,8 +1,7 @@
 ---
 title: Custom Objects
-sidebar_position: 1
-sidebar_custom_props:
-  icon: TbAugmentedReality
+position: 1
+icon: TbAugmentedReality
 ---
