+++ packages/twenty-website/src/content/user-guide/basics/notes.mdx
@@ -1,11 +1,9 @@
 ---
 title: Notes
-sidebar_position: 1
-sidebar_custom_props:
-  icon: TbNote
+position: 1
+icon: TbNote
 ---
 
-import PostImage from '@theme/PostImage';
 
 Easily create a note to keep track of important information. 
 
@@ -13,7 +11,7 @@ Easily create a note to keep track of important information.
 
 To attach a note to the record, go to the <b>Notes</b> tab of a record page and click on `+ New Note`. You can also format, comment, and upload images to your notes.   
 
-<img src="images/user-guide/create-new-note-light.png" style={{width:'100%', maxWidth:'800px'}}/>
+<img src="/images/user-guide/create-new-note-light.png" style={{width:'100%', maxWidth:'800px'}}/>
 
 ## Format Notes
