+++ packages/twenty-website/src/content/user-guide/basics/opportunities.mdx
@@ -1,11 +1,8 @@
 ---
 title: Opportunities
-sidebar_position: 1
-sidebar_custom_props:
-  icon: TbTargetArrow
+position: 1
+icon: TbTargetArrow
 ---
-
-import PostImage from '@theme/PostImage';
 
 All opportunities are presented in a Kanban board, where each column represents the stage of your workflow and each card represents a record. For each card, you can list the amount, close date, probability, and the point of contact. You can also move each card between stages as it goes through your workflow.
