+++ packages/twenty-website/src/content/user-guide/basics/tasks.mdx
@@ -1,8 +1,7 @@
 ---
 title: Tasks
-sidebar_position: 1
-sidebar_custom_props:
-  icon: TbChecklist
+position: 1
+icon: TbChecklist
 ---
 
 import PostImage from '../../../components/PostImage'
