+++ packages/twenty-website/src/content/user-guide/home.mdx
@@ -1,11 +1,7 @@
 ---
 title: Get started
-displayed_sidebar: userSidebar
-sidebar_class_name: hidden
-sidebar_position: 0
-sidebar_custom_props:
-  icon: TbUsers
-  isSidebarRoot: true
+position: 0
+icon: TbUsers
 ---
