+++ packages/twenty-website/src/content/user-guide/integrations/connect-zapier.mdx
@@ -1,8 +1,7 @@
 ---
-title: Connect Twenty and Zapier
-sidebar_position: 3
-sidebar_custom_props:
-  icon: TbBrandZapier
+title: Connect Zapier
+position: 3
+icon: TbBrandZapier
 ---
 
 :::caution
@@ -21,7 +20,7 @@ Sync Twenty with 3000+ apps using [Zapier](https://zapier.com/), and automate yo
 6. Enter your API key and click on 'Yes, Continue to Twenty.'
 
 <div style={{textAlign: 'center'}}>
-    <img src="/images/user-guide/connect-zapier.png" alt="Connect Twenty to Zapier" />
+    <img src="/images/user-guide/connect-zapier.png" alt="Connect Twenty to Zapier" style={{width:'100%', maxWidth:'800px'}}/>
 </div>
 
 You can now continue creating your automation!
