+++ packages/twenty-website/src/content/user-guide/integrations/generating-api-keys.mdx
@@ -1,8 +1,7 @@
 ---
 title: Generating an API Key
-sidebar_position: 2
-sidebar_custom_props:
-  icon: TbApi
+position: 2
+icon: TbApi
 ---
 
 To generate an API key:
