+++ packages/twenty-website/src/content/user-guide/others/glossary.mdx
@@ -1,8 +1,7 @@
 ---
 title: Glossary
-sidebar_position: 3
-sidebar_custom_props:
-  icon: TbVocabulary
+position: 3
+icon: TbVocabulary
 ---
 
 ### Company & People
