+++ packages/twenty-website/src/content/user-guide/others/tips.mdx
@@ -1,8 +1,7 @@
 ---
 title: Tips 
 sidebar_position: 1
-sidebar_custom_props:
-  icon: TbInfoCircle
+icon: TbInfoCircle
 ---
 
 ## Update workspace name & logo
